Apply a schema change by running it through a state machine
A schema change is now run through many stages, where each stage runs
as a transaction on the database. Some steps involve the use of
LeaseManager.Publish(), that publishes a new version to the schema
only after there are only outstanding leases to the existing version of
the schema.

Since the synchronous execution of a schema change can
fail at any time, a second asynchronous path is added to
move the state machine along.